requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
arrow==1.3.0
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
av==10.0.0
backoff==2.2.1
beautifulsoup4==4.12.2
blessed==1.20.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cmaes==0.10.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.7.0
contourpy==1.1.1
croniter==1.4.1
ctranslate2==3.20.0
cycler==0.12.0
dateutils==0.6.12
decorator==5.1.1
deepdiff==6.5.0
docopt==0.6.2
einops==0.7.0
exceptiongroup==1.1.3
fastapi==0.103.2
faster-whisper==0.9.0
ffmpeg-python==0.2.0
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.43.0
frozenlist==1.4.0
fsspec==2023.9.2
future==0.18.3
h11==0.14.0
huggingface-hub==0.16.4
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.4
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
julius==0.2.7
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
lightning==2.0.9.post0
lightning-cloud==0.5.39
lightning-utilities==0.9.0
llvmlite==0.41.0
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
networkx==3.1
nltk==3.8.1
numba==0.58.0
numpy==1.25.2
omegaconf==2.3.0
onnxruntime==1.16.0
optuna==3.3.0
ordered-set==4.1.0
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
platformdirs==3.11.0
pooch==1.7.0
primePy==1.3
protobuf==4.24.4
psutil==5.9.5
pyannote.audio @ git+https://github.com/pyannote/pyannote-audio@db24eb6c60a26804b1f07a6c2b39055716beb852
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.9.post0
pytorch-metric-learning==2.3.0
pytz==2023.3.post1
PyYAML==6.0.1
readchar==4.0.5
regex==2023.10.3
requests==2.31.0
rich==13.6.0
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.3
semver==3.0.1
sentencepiece==0.1.99
shellingham==1.5.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.6
speechbrain==0.5.15
SQLAlchemy==2.0.21
starlette==0.27.0
starsessions==1.3.0
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tokenizers==0.14.0
torch==2.0.0
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==2.0.1
torchmetrics==1.2.0
torchvision==0.15.1
tqdm==4.66.1
traitlets==5.11.2
transformers==4.34.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
uvicorn==0.23.2
wcwidth==0.2.8
websocket-client==1.6.3
websockets==11.0.3
whisperx @ git+https://github.com/m-bain/whisperx.git@07fafa37b3ef7ce8628b194da302a5a996bb7d37
yarl==1.9.2
